Create README.md
Implementation of Viterbi algorithm, which is a dynamic algorithm, used to determine the most likely sequence of hidden states.  In this implementation it works on a HMM model, here its used to determine the sequence of variables which is the underlying source of some sequence of observations,  this is a  decoding  task.